feat: moved privileges to its own tab in ACP
