+++ public/language/en-GB/admin/manage/privileges.json
@@ -35,6 +35,7 @@
 	"acp.general": "General",
 	"acp.manage": "Manage",
 	"acp.settings": "Settings",
+	"acp.privileges": "Privileges",
 	"acp.appearance": "Appearance",
 	"acp.extend": "Extend",
 	"acp.advanced": "Advanced"
