+++ src/privileges/admin.js
@@ -15,6 +15,7 @@ module.exports = function (privileges) {
 		{ name: '[[admin/manage/privileges:acp.general]]' },
 		{ name: '[[admin/manage/privileges:acp.manage]]' },
 		{ name: '[[admin/manage/privileges:acp.settings]]' },
+		{ name: '[[admin/manage/privileges:acp.privileges]]' },
 		{ name: '[[admin/manage/privileges:acp.appearance]]' },
 		{ name: '[[admin/manage/privileges:acp.extend]]' },
 		{ name: '[[admin/manage/privileges:acp.advanced]]' },
@@ -24,6 +25,7 @@ module.exports = function (privileges) {
 		'acp:general',
 		'acp:manage',
 		'acp:settings',
+		'acp:privileges',
 		'acp:appearance',
 		'acp:extend',
 		'acp:advanced',
