+++ src/routes/admin.js
@@ -56,7 +56,6 @@ function addRoutes(router, middleware, controllers) {
 	router.get('/manage/categories/:category_id', middlewares, controllers.admin.categories.get);
 	router.get('/manage/categories/:category_id/analytics', middlewares, controllers.admin.categories.getAnalytics);
 
-	router.get('/manage/privileges/:cid?', middlewares, controllers.admin.privileges.get);
 	router.get('/manage/tags', middlewares, controllers.admin.tags.get);
 	router.get('/manage/post-queue', middlewares, controllers.admin.postQueue.get);
 	router.get('/manage/ip-blacklist', middlewares, controllers.admin.blacklist.get);
@@ -81,6 +80,8 @@ function addRoutes(router, middleware, controllers) {
 	router.get('/manage/uploads', middlewares, controllers.admin.uploads.get);
 
 	router.get('/settings/:term?', middlewares, controllers.admin.settings.get);
+
+	router.get('/privileges/:cid?', middlewares, controllers.admin.privileges.get);
 
 	router.get('/appearance/:term?', middlewares, controllers.admin.appearance.get);
