+++ src/socket.io/admin.js
@@ -34,6 +34,7 @@ var SocketAdmin = {
 	navigation: require('./admin/navigation'),
 	rooms: require('./admin/rooms'),
 	social: require('./admin/social'),
+	privileges: require('./admin/privileges'),
 	themes: {},
 	plugins: {},
 	widgets: {},
@@ -50,6 +51,7 @@ var SocketAdmin = {
 		general: new Set(['analytics', 'rooms', 'navigation', 'social']),
 		manage: new Set(['user', 'categories', 'groups', 'tags', 'uploads']),
 		settings: new Set(['rewards', 'widgets', 'config', 'settings', 'email']),
+		privileges: new Set(['privileges']),
 		appearance: new Set(['themes']),
 		extend: new Set(['plugins']),
 		advanced: new Set(['logs', 'errors']),
