+++ src/views/admin/privileges.tpl
